requirement.txt
PyAudio
SpeechRecognition
pydub
langchain
langchain-google-genai
python-dotenv
pyttsx3
streamlit